{"types":"C","runs":44}

+++ Practice19.java
@@ -4,7 +4,7 @@ import acm.program.ConsoleProgram;
 public class Practice19 extends ConsoleProgram {
 	public void run(){
 		GLabel lab = new GLabel ("dasdadad", 100, 110);
-		double a = lab.getWidth();
+		int a = (int) lab.getWidth();
 		print(lab);
 	}
 }
